remove Auth and added Double support to DecimalPicker

+++ Sources/Ra9rCore/Extensions/Double+Extenstion.swift
@@ -1,13 +1,13 @@
 //
-//  Int+Extension.swift
-//  Ra9rKit
+//  File.swift
+//  
 //
-//  Created by Rodney Aiglstorfer on 11/23/23.
+//  Created by Rodney Aiglstorfer on 12/31/23.
 //
 
 import Foundation
 
-extension Float : Uncombinable {
+extension Double : Uncombinable {
     
     /// Formats the into a `String` with a specified number of decimals
     public func formatPrecision(_ decimals: Int) -> String {
@@ -16,13 +16,13 @@ extension Float : Uncombinable {
     
     /// Separates a `Float` into its `wholeNumber` and `decimalNumber` components
     public func uncombine() -> (wholeNumber: Int, decimalNumber: Int) {
-        (wholeNumber: Int(self), decimalNumber: Int( ((self - Float(Int(self))) * 10)) )
+        (wholeNumber: Int(self), decimalNumber: Int( ((self - Double(Int(self))) * 10)) )
     }
     
     /// An internal function that will produce a float value that is the combination of a
     /// wholenumber and a decimal number from the picker
     public init(_ wholeNumber: Int, _ decimalNumber: Int) {
-        self = Float(wholeNumber) + (Float(decimalNumber) * 0.1)
+        self = Double(wholeNumber) + (Double(decimalNumber) * 0.1)
     }
     
 }
